modifications to make openocd reliable with betrusted
I tried to give them a patch that makes this thing actually
work, but of course, they never took it and they took someone
else's, which of course misses a major corner case. So it works
fine for things like configuring an FPGA, but if you try burning
a ROM with it, the bit error rate is non-negligible and things
are sad.

Yet another openOCD fork.